refactor(personal-assistant): extract definitions, goals, workflows, status domains
Phase B complete: all 26 LifeOps domains are now sub-services. definitions/goals/
workflows/status inject the reminders-hub methods they consume (typed deps).
status reads the 4 ex-collision methods (getBrowserSettings, listBrowserCompanions,
getHealthConnectorStatus, getXConnectorStatus) via typed deps — a compat
StatusMixinDependencies export keeps service.ts compiling until the Phase-C
facade drops the Omit<>s. workflows keeps runDueWorkflows/runDueEventWorkflows as
public delegators (consumed by the reminders scheduler).

0 errors in src/lifeops/ scope; tests 105 pass / 4 pre-existing fails. No
behaviour change.
